Move analysis into analyses, clean up the resulting fallout
