+++ angr/analyses/boyscout.py
@@ -1,10 +1,9 @@
-import re
 import logging
+import re
 from collections import defaultdict
 
+from . import Analysis, register_analysis
 from archinfo import all_arches
-
-from ..analysis import Analysis, register_analysis
 
 l = logging.getLogger("angr.analyses.boyscout")
