+++ angr/project.py
@@ -707,7 +707,7 @@ from .errors import AngrError
 from .factory import AngrObjectFactory
 from .simos import SimOS, os_mapping
 from .extern_obj import AngrExternObject
-from .analysis import Analyses
+from .analyses.analysis import Analyses
 from .surveyor import Surveyors
 from .knowledge_base import KnowledgeBase
 from .engines import SimEngineFailure, SimEngineSyscall, SimEngineHook
